refactor: change constants import
